requirements.txt
# AI/ML Core
langchain==0.1.0
langchain-community==0.0.10
langchain-groq==0.0.1
groq==0.4.0

# Vector Database & Embeddings
chromadb>=0.4.15
langchain-chroma>=0.1.0
langchain-huggingface>=0.0.1
sentence-transformers==2.2.2

# PDF Processing (for knowledge base)
pypdf2>=3.0.0

# Computer Vision & Disease Detection
pillow==10.0.0
numpy==1.24.3

# Geographic Analysis
geopy==2.3.0

# Web Frameworks
fastapi==0.104.0
uvicorn==0.24.0
streamlit==1.28.0

# WhatsApp Integration
twilio==8.9.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6
